LayoutTests/tiled-drawing/scrolling/scroll-snap/scroll-snap-proximity-overflow-expected.txt

Scrolling in container with 2 drag ticks
- Did the scrolling snap to the top? YES
- Did scrolling snap to the second box? NO
Scrolling in container with 31 drag ticks
- Did the scrolling snap to the top? NO
- Did scrolling snap to the second box? NO
Scrolling in container with 59 drag ticks
- Did the scrolling snap to the top? NO
- Did scrolling snap to the second box? YES
